+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/rizkybiz/terraform-provider-hashicorpversions
+module github.com/rizkybiz/terraform-provider-hashicorpproducts
 
 go 1.17
 
@@ -9,7 +9,6 @@ require (
 
 require (
 	github.com/agext/levenshtein v1.2.2 // indirect
-	github.com/apparentlymart/go-cidr v1.0.1 // indirect
 	github.com/apparentlymart/go-textseg/v12 v12.0.0 // indirect
 	github.com/apparentlymart/go-textseg/v13 v13.0.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
