+++ hashicorpproducts/provider.go
@@ -1,0 +1,14 @@
+package hashicorpproducts
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// Provider -
+func Provider() *schema.Provider {
+	return &schema.Provider{
+		DataSourcesMap: map[string]*schema.Resource{
+			"hashicorp_products_version": dataSourceVersion(),
+		},
+	}
+}
