+++ hashicorpproducts/provider_test.go
@@ -1,0 +1,17 @@
+package hashicorpversions
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+func TestProvider(t *testing.T) {
+	if err := Provider().InternalValidate(); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+}
+
+var testProviders = map[string]*schema.Provider{
+	"hashicorp_versions": Provider(),
+}
