--- hashicorpversions/provider.go
@@ -1,14 +1,0 @@
-package hashicorpversions
-
-import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-)
-
-// Provider -
-func Provider() *schema.Provider {
-	return &schema.Provider{
-		DataSourcesMap: map[string]*schema.Resource{
-			"hashicorp_versions": dataSourceVersion(),
-		},
-	}
-}
