+++ main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
-	"github.com/rizkybiz/terraform-provider-hashicorpversions/hashicorpversions"
+	"github.com/rizkybiz/terraform-provider-hashicorpversions/hashicorpproducts"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
-			return hashicorpversions.Provider()
+			return hashicorpproducts.Provider()
 		},
 	})
 }
